Refactored z-index logic to separate file
